Piece system changes and improvements
Changes:
- Instead of using a class for parts, it now uses a virtual class.*
- Added a little information about which parts are played.
- Moved "vector2" structure to Piece.h (It will have its own file later.)
- Changed "convert-vector2" function.**
- I got rid of the "PIECES_MATRIX" macro. Yes, it was a really tough decision.
- Get rid of the gap between engine and start function in "main.cpp".

* There is a different class for each part derived from the virtual class.
In this way, the movements of a selected part can be controlled with a function in its class.
So, we got rid of using a lot of if-else.

** Because the definition of piece position in chess game is very different from
vector coordinates and cartesian coordinates in cpp. I was caught off guard by a big bug.
Now every position value is stored according to the way cpp works.

You don't know but after making this distinction the code stopped working.
I changed the code a lot but I didn't upload it here because even the pieces can't be moved.
Thanks to "Anılcan Günkaya", he is now working. And I have the courage to continue this.